requirements.txt
click==6.7
Flask>=0.12.3
Flask-Cors==3.0.3
itsdangerous==0.24
Jinja2>=2.10.1
MarkupSafe==1.0
six==1.11.0
Werkzeug==0.15
flask-mongoengine
google-cloud-storage
google-api-python-client